# openapi.yaml
openapi: 3.1.0
info:
  title: LazyHR API
  description: |
    Human Resources Management System API
    
    **Note**: This API has some path ambiguity between `/api/leave/user/{userId}` and `/api/leave/{leaveId}/*` endpoints.
    In practice, this works because:
    - `user` is a literal string and not interpreted as a leaveId
    - The Spring Boot framework resolves these paths correctly based on the request method and path specificity
    
    However, OpenAPI linting tools may flag these as ambiguous.
  version: 1.0.0
  contact:
    name: LazyHR Team
    email: support@lazyhr.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/lazyhr
    description: Production server
  - url: http://localhost:8080/lazyhr
    description: Development server
...
paths:
...
  /api/attendance/clock-in:
    post:
      tags:
        - Attendance
      summary: Clock in
      description: Record employee clock in time
      operationId: clockIn
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully clocked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Clock in failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/clock-out:
    post:
      tags:
        - Attendance
      summary: Clock out
      description: Record employee clock out time
      operationId: clockOut
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully clocked out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Clock out failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/today/{userId}:
    get:
      tags:
        - Attendance
      summary: Get today's attendance
      description: Get today's attendance record for a specific user
      operationId: getTodayAttendance
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Today's attendance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Attendance'
        '400':
          description: Failed to fetch attendance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/user/{userId}:
    get:
      tags:
        - Attendance
      summary: Get user attendance by date range
      description: Get attendance records for a user within a date range
      operationId: getUserAttendance
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: startTimestamp
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Start date as Unix timestamp in milliseconds
          example: 1721865600000
        - name: endTimestamp
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: End date as Unix timestamp in milliseconds
          example: 1722038400000
      responses:
        '200':
          description: Attendance records retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch attendance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/today:
    get:
      tags:
        - Attendance
      summary: Get all attendance for today
      description: Get today's attendance records for all users
      operationId: getTodayAttendances
      responses:
        '200':
          description: Today's attendance records retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch today's attendances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/history/{userId}:
    get:
      tags:
        - Attendance
      summary: Get user's attendance history
      description: Get complete attendance history for a user
      operationId: getUserAttendanceHistory
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attendance history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch attendance history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/status/{userId}:
    get:
      tags:
        - Attendance
      summary: Check if user is currently clocked in
      description: Get current attendance status for a user
      operationId: getAttendanceStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attendance status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AttendanceStatus'
        '400':
          description: Failed to fetch attendance status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/{attendanceId}/notes:
    put:
      tags:
        - Attendance
      summary: Update attendance notes
      description: Update notes for an attendance record
      operationId: updateAttendanceNotes
      parameters:
        - name: attendanceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              description: Notes to update
      responses:
        '200':
          description: Notes updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to update notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/{attendanceId}/break:
    put:
      tags:
        - Attendance
      summary: Update break duration
      description: Update break duration for an attendance record
      operationId: updateBreakDuration
      parameters:
        - name: attendanceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: breakMinutes
          in: query
          required: true
          schema:
            type: integer
          description: Break duration in minutes
      responses:
        '200':
          description: Break duration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to update break duration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/attendance/stats/today:
    get:
      tags:
        - Attendance
      summary: Get attendance statistics
      description: Get today's attendance statistics
      operationId: getTodayStats
      responses:
        '200':
          description: Today's attendance count retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: Response status
        message:
          type: string
          description: Response message
        data:
          type: object
          description: Response data
      required:
        - status
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        hireDate:
          type: integer
          format: int64
          description: Date of hire (Unix timestamp in milliseconds)
          example: 1609459200000
          nullable: true
        salary:
          type: number
          format: double
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
        createdAt:
          type: integer
          format: int64
          description: User creation timestamp (Unix timestamp in milliseconds)
          example: 1709251200000
        updatedAt:
          type: integer
          format: int64
          description: User last update timestamp (Unix timestamp in milliseconds)
          example: 1721779200000
    Attendance:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        attendanceDate:
          type: integer
          format: int64
          description: Attendance date (Unix timestamp in milliseconds)
          example: 1721865600000
        clockInTime:
          type: integer
          format: int64
          description: Clock in time (Unix timestamp in milliseconds)
          example: 1721822400000
          nullable: true
        clockOutTime:
          type: integer
          format: int64
          description: Clock out time (Unix timestamp in milliseconds)
          example: 1721851200000
          nullable: true
        totalHours:
          type: number
          format: double
        status:
          type: string
          enum: [PRESENT, ABSENT, LATE, HALF_DAY]
...
    Role:
      type: string
      enum:
        - ADMIN
        - MANAGER
        - EMPLOYEE
      description: User role in the system
...
    AttendanceStatus:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User ID
        isClockedIn:
          type: boolean
          description: Whether the user is currently clocked in
        activeAttendance:
          $ref: '#/components/schemas/Attendance'
          nullable: true
          description: Current active attendance record if clocked in
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []